Move metrics types to DataLayer domain
