+++ js/src/domains/interventions/components/TargetPopulationForm.tsx
@@ -1,8 +1,8 @@
 import React, { FC, useCallback, useMemo } from 'react';
 import { Stack } from '@mui/material';
 import InputComponent from 'Iaso/components/forms/InputComponent';
+import { MetricType } from '../../dataLayers/types/metrics';
 import { MESSAGES } from '../../messages';
-import { MetricType } from '../../planning/types/metrics';
 
 type Props = {
     targetPopulation: string[];
